+++ src/scanner/enumerate.rs
@@ -144,6 +144,7 @@ pub fn enumerate_filesystem_inputs(
         Some(&matcher_stats),
         enable_profiling,
         Some(shared_profiler),
+        args.compat_ignore_comments,
     )?;
     let blob_processor_init_time = Mutex::new(t1.elapsed());
     let make_blob_processor = || -> BlobProcessor {