+++ tests/int_github.rs
@@ -118,6 +118,7 @@ fn test_github_remote_scan() -> Result<()> {
             since_commit: None,
             branch: None,
         },
+        compat_ignore_comments: false,
         content_filtering_args: ContentFilteringArgs {
             max_file_size_mb: 25.0,
             no_extract_archives: false,